requirements.txt
asgiref==3.3.1
attrs==20.3.0
coverage==5.5
Django==3.1.7
djangorestframework==3.12.2
drf-spectacular==0.14.0
inflection==0.5.1
jsonschema==3.2.0
psycopg2-binary==2.8.6
pyrsistent==0.17.3
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
